inference/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
pillow
numpy
torch
torchvision
torchaudio
ultralyticsplus
transformers
accelerate
timm